Modifying the bazel install and adding bash completion in Dockerfile to enable tab completion (#15)
## Description of change
To achieve tab completion functionality in Bazel, we needed to modify the Docker file to first install bash-completion and enable it, then install Bazel via the apt repository. This also installs the bazel completion scripts that Bazelisk unfortunately does not.

## Asana task link
[Asana task](https://app.asana.com/0/1202178773526279/1202206229790774/f)
## Guide to reproduce test results.
Build and run the docker container. 
Enter the source directory and start typing a bazel command, then press tab.
## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [ ] This change is covered by tests that are already landed, or in this PR.