Add codecov.yml team config and set the comment behavior to post new comments (#62)
## Description of change
The current codecov behavior (default) is set to post a single codecov comment and edit it whenever there is a new report. This behavior results in a single comment that sometimes lags behind what appears in the GH actions report and is confusing. This new change will change the behavior to post a new comment every time there is a new report and delete the old report. 
## Asana task link
https://app.asana.com/0/1202178773526279/1202416205185944/f
## Guide to reproduce test results.
Push a code change and view report multiple times. 

See example comment of a changing PR in this comment: 
![image](https://user-images.githubusercontent.com/104402278/172940956-1d289caa-b190-4ecd-8a67-0082e671a831.png)

## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [ ] This change is covered by tests that are already landed, or in this PR.