Pinning bazel version to avoid build breakages. (#71)
## Description of change
Our repo asserts the expected bazel version (in .bazelversion). However, our
Dockerfile apt installs bazel from the bazel repository, which can update the
version. Looks like it just updated to 5.2, which causes a global breakage next
time the container is rebuilt. The right solution is probably to pin the bazel
version, which this PR does.

## Asana task link
[Asana task](https://app.asana.com/0/1202178773526279/1202433926737350/f)
## Guide to reproduce test results.
do:
```
dev/build.sh
dev/run.sh
```
Then issue any bazel command and observe that is does not explode.
To test the failure case:
Make any trivial change to the Dockerfile that is no this one and repeat the
test above. Observe that is does explode.

## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [ ] This change is covered by tests that are already landed, or in this PR.