Exclude easily swappable parameters check. (#81)
## Description of change
"bugprone-easily-swappable-parameters" is not the most desirable clang rule.
Mainly because supporting it requires the user to create custom structs to hold
the data, but initializing the structs has the same problem this rule claims to
solve.

## Asana task link
[Asana task](https://app.asana.com/0/1202178773526279/1202491914495186/f)
## Guide to reproduce test results.
Patch this change onto a PR that is currently failing with this issue 
(e.g. #73) and verify that this silences the error.
## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in this PR.