Check macro & Codecov issue workaround (#94)
## Description of change
This change addresses a strange issue where if you line break at the streaming
operator into a CHECK() macro then Codecov treats it as a separate line and 
believes it has no code coverage. The only currently know workaround is simply 
not to do this. This PR removes one such line-break that is erroneously pulling
our coverage metric down.
## Asana task link
[Asana task](https://app.asana.com/0/1202178773526279/1202672342571867/f)
## Guide to reproduce test results.
Check the coverage report for the PR.
## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in this PR.