Additional adjoint methods for SO3, SE3 (#93)
## Description of change
Adding additional methods to S03 and SE3 to support the adjoint operator on
both the group and the algebra. This includes forwarding additions to framed
group to support substitution.

These methods are needed to support the forthcoming time parameterized curve.
The design doc for this is linked via the Asana task.
## Asana task link
[Asana task](https://app.asana.com/0/0/1202672342571863/f)
## Guide to reproduce test results.
```
bazel test //transforms/...
```
## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in this PR.