Quintic Polynomical Coefficients helper (#100)
## Description of change
Compute the polynomial coefficients for a quintic Hermite spline.
Polynomial coefficients calculation is straightforward, but requires a few
magic numbers. This is unavoidable, but to save the reader of the Hermite spline
class from wading through this clutter, we are abstracting these out to a helper
class that can be called to serve up the required coefficients.

We can test this class for correctness against hand computed coefficients, which
also helps build up incremental testing of the components of the Hermite spline
## Asana task link
[Asana task](https://app.asana.com/0/0/1202825153541987/f)
## Guide to reproduce test results.
```
bazel test //curves:quintic_poly_test
```
## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in this PR.