Setting up basic docs build and deploy (#140)
## Description of change
Setting up a basic implementation of docs creation, with automated build and 
deploy. The docs are built using `mkdocs` with both local and web deployment
options. The PR includes a github action to deploy doc changes to the official 
docs site on merge.
## Asana task link
[Asana task](https://app.asana.com/0/1202178773526279/1203248753995779/f)
## Guide to reproduce test results.
Testing locally:
Make changes to the docs run `mkdocs serve` within the docker container. 
Visit the URL printed in a browser on the machine running the container.
Testing web deployment.
If the workflow file is changed from `on push` to `on pull-request` then the 
docs will build on changes pushed to this PR and deploy to the site.
## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [ ] This change is covered by tests that are already landed, or in this PR.