Add trunk to docker container and initialize it in the repo (#152)
## Description of change
This change adds the default configurations for Trunk to the repository by running `trunk init`. 

This change installs trunk in the docker image. 

This change does not expect to change our linting GH Action or any other behavior. 

## Asana task link
https://app.asana.com/0/1202178773526279/1203353731322520/f
## Guide to reproduce test results.
Build the docker container and validate that `trunk --version` returns a valid response. 

## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [ ] This change is covered by tests that are already landed, or in this PR.