Trajectory class (#162)
## Description of change

Adding a Trajectory class that wraps TCurve and works with RigidBodyStates. 
While the TCurve represents an abstract spline, a Trajectory is designed to 
explicitly describe the motion of a six-dof rigid body over time. Thus it sets
and gets RigidBodyStates rather than raw TwoJets and works with time::timestamps
for ease of association with actors.

## Asana task link
[Asana task](https://app.asana.com/0/1202178773526279/1203608723315727/f)

## Guide to reproduce test results.
```
bazel test //resim_core/actor/state:trajectory_test
```
## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in this PR.