Refactor TwoJet -> TwoJetL (#190)
## Description of change
In PR (173) we changed TwoJet to have two flavors TwoJetL and TwoJetR. This 
created a lot of places where The old name TwoJet was used for what is now
called TwoJetL this refactor changes the name throughout the codebase and 
removes the alias that supported using both names.
## Asana task link
[Asana task](https://app.asana.com/0/1202178773526279/1203608723315721/f)
## Guide to reproduce test results.
This touches many libraries. The easies test is to run all tests.
```
bazel test ...
```
## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in this PR.