TCurve Packers and Unpackers (#268)
## Description of change

Adds packers and unpackers for TCurves to and from protobuf. 

It follows the pattern of introducing a protobuf message type for each possible flavour of TCurve, each with the same field names and nested message. 

Packing and unpacking by and large calls the component packers, so most code in this PR is boilerplate.

## Asana task link

[Asana task](https://app.asana.com/0/1203751014069901/1203869115126127/f)

## Guide to reproduce test results.

```
bazel test //resim_core/curves/proto:t_curve_to_proto_test
```

## Checklist:
- [X] I have self-reviewed this change.
- [X] I have tested this change.
- [X] This change is covered by tests that are already landed, or in this PR.