Creates a TCurve Constructor for Trajectory (#269)
## Description of change

Adds a TCurve Constructor for Trajectory and constrains TCurve points to have consistent point frames.

The original implementation of a TCurve enforced reference frame consistency between points (for sanity),
but was agnostic about the frame of individual points. The concrete downstream use of TCurve that we currently
have, a trajectory, requires the point frame (body frame in the context of a trajectory) to be consistent.

Thus, to be consistent with this, and in absence of other use cases right now, we have taken the decision to
constrain the TCurve itself to consistent point frames. 

We are fully open to relaxing this constraint in future if a use case comes up.

## Asana task link

[Asana task](https://app.asana.com/0/1202178773526279/1203638018487834/f)

## Guide to reproduce test results.

```
bazel test //resim_core/curves:t_curve_test
bazel test //resim_core/trajectory:trajectory_test
```

## Checklist:
- [X] I have self-reviewed this change.
- [X] I have tested this change.
- [X] This change is covered by tests that are already landed, or in this PR.