Framed Vectors (#325)
## Description of change
When summing many forces or speeds (e.g. wind velocity) in different frames, it
can be useful to do some bookkeeping on which frame each vector is expressed in.
We can extend our use of FramedGroup to also support FramedVector, which allows
us to be explicit about the Frame in which a vector is expressed and do frame
checking on summing and transform operations.
## Asana task link
[Asana task](https://app.asana.com/0/0/1204267373713142/f)
## Guide to reproduce test results.
```
bazel test //resim-core/transforms:framed_vector_test
bazel test //resim-core/transforms:framed_group_test
```
## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in this PR.