Make SO3 and SE3 Framed by default (#376)
This change makes SO3 and SE3 framed by default. 
## Description of change
Unframed transforms are still possible by simply not setting the frames in SO3, 
SE3 or by setting them to null with the new `.set_unframed()` method. Behavior 
is equivalent to `FramedGroup<S*3>` and all FramedGroup test are ported to 
`liegroup_test.cc` in this PR.

This change makes `FramedGroup<S*3>` obsolete, but does not remove it from the 
repo to limit the size of the change. It should be safe to land this change 
regardless. However there should be a plan in place to remove `FramedGroup<S*3>`
fairly quickly thereafter to avoid divergence. This plan would be approximately:

1. Extend `proto:S*3` to include frames (making it equivalent to `proto:FS*3`).
2. Working back from the leaf nodes in the bazel tree, remove 
   `FramedGroup<S*3>` from one lib at a time.

## Asana task link
[Asana task](https://app.asana.com/0/0/1204729511847792/f)
## Guide to reproduce test results.
```
bazel test //resim_core/transforms/...
```
## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in this PR.